Update VM build
build VM to adapt to new ci/cd workflow.
The VM image is built on gerrit verify when the image is not present in google storage.
It is not uploaded yet on google storage.
The build + upload of the new VM image is done after the review is merged.
For details on how this is done, refer to ./jjb/nfvbench/nfvbench.yaml in the opnfv releng repository.

Change-Id: I147d4ee6d101e9f2a7cf55fadd637c7fc89467a0
Signed-off-by: ahothan <ahothan@cisco.com>