Update patch set 1
Patch Set 1:

Just a general comment on this enhancement request.
Currently the number of flows requested translates into as many dest IP addresses, all other factors of the tuple being constant.
Adding random values in UDP src/dst ports will clearly have an impact on the number of actual flows since the number of tuples will increase in a way that is not completely deterministic.

So adding this option will clearly inflate the actual number of flows generated.
I think that should be made clear in the doc.

Patch-set: 1