Update patch set 1
Patch Set 1:

(2 comments)

Overall looks ok although the vxlan/mpls feature is untested.
General comment is that pps calculation could have been approximative, we can assume a few pps deviation is not unreasonable when GARP Is enabled, but the proposed diffs seems to get this properly.
The more critical part is related to exact packet count and drops which is important for NDR/ZDR benchmarks (zero drop rate), this will require testing and make sure ZDR calculation are not off due to the added TX GARP packet. To be discussed in more detail in slack

Patch-set: 1
CC: Gerrit User 5807 <5807@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>