behave_tests: increase nfvbench_test_api timeout
In some cases, the 50 seconds timeout to wait for nfvbench HTTP server to be
ready is too short.  (Convoluted but real example: when the DNS servers are not
properly configured and nfvbench tries to reach the OpenStack APIs while it
does not need them because we just want to do a loopback test without loop VM)

The new timeout is 120 seconds.

Change-Id: I4932eff7c9a100370e7ceaaa2a467efbbceb5993
Signed-off-by: Gwenael Lambrouin <gwenael.lambrouin@orange.com>