behave_tests: configure nfvbench ip/port with env vars
It is now possible to configure nfvbench server IP address and port number
with environment variables: NFVBENCH_SERVER_HOST and NFVBENCH_SERVER_PORT.

It is still possible to configure them in feature files, and the values found
in feature files take precedence.

This allows to have behave tests and nfvbench server running on different
machines without changing feature files, which is especially useful for
testing.

Signed-off-by: Gwenael Lambrouin <gwenael.lambrouin@orange.com>
Change-Id: I98dc7f87a1a233b90b44dfc8b26a1e63961fff3c