behave_tests: change packet rate for latency non regression tests
Base the rate of the latency test on the latest characterization max throughput
test instead of the latest non regression max throughput test.  The goal is to
use the same packet rate for all latency tests and to avoid variations of the
latency result due to the variation of the max throughput results, ie to
decouple max throughput and latency test results.

This is achieved with a new "Given" behave phrase:

    Given packet rate equal to {percentage} of max throughput of last characterization

This new phrase is now used by default in non-regression.feature, but it is
still possible to use the previous behaviour with the phrase:

    Given <throughput> rate of previous scenario

Change-Id: I15b5d7a68cd57c67d01d2119781f65114e6d41ce
Signed-off-by: Gwenael Lambrouin <gwenael.lambrouin@orange.com>