Update patch set 2
Patch Set 2:

Thanks for this feedback. I guess we can drop GS_URL and WORKSPACE because we don't use them for nfvbench: this will improve clarity in .gitlab-ci.yml.

Another question: would you like to review systematically all the patches related to gitlab ci? If so, could you cast your vote on the patchset? Or do you consider that after your initial contribution it is now up to the projet to develop and maintain the .gitlab-ci.yml files?

Patch-set: 2