nfvbenchvm: set mgmt interface MTU=1500 by default
By default, configure the MTU of the management interface to the
conservative value of 1500: this will reduce the risk to get an
unmanageable VM.  Else the MTU we will get will depend on the IaaS setup
and it could be high.  And this will lead to troubles when we try to
manage the VM from a machine configured with a high MTU and when the
network path contains a link with a low MTU.  This can typically happen
when we try to manage the VM from another VM in the same IaaS and when
the management traffic has to go outside the IaaS.

To set the MTU to a different value, configure the INTF_MGMT_MTU
variable in /etc/nfvbenchvm.conf.  If INTF_MGMT_MTU is set to the
special value "auto", the MTU will not be configured and it will keep
the value set by the hypervisor ("legacy" nfvbenchvm behavior).  If
INTF_MGMT_MTU is unset, the MTU will be set to 1500.  In other cases,
the MTU will be set to the value of INTF_MGMT_MTU.

Change-Id: I5816113bbacec5fcdcc9672e20d8a4e2ac0c599a
Signed-off-by: Gwenael Lambrouin <gwenael.lambrouin@orange.com>