nfvbenchvm: abort build on error
Make sure a VM image build fails if any step fails.  Else we can end up
with a bad image not containing all that we want, and we discover this
later at run time.

For instance, without the current patch, we can get a generator VM image
without nfvbench installed because "pip install" failed, but this is not
detected by the build process because the last step in
post-install.d/52-nfvbench-script is an "echo" command that returns a
success code.

We fix this by making all the Bash scripts fail if any command whose
return code is not tested fails (set -euo pipefail).

Change-Id: Ic5ec9eb85a8d6e7e4d1dcbac8ebcac5931e4808e
Signed-off-by: Gwenael Lambrouin <gwenael.lambrouin@orange.com>