nfvbenchvm: add option to use local nfvbench code
build-image.sh has a new option -s to build the generator VM image from
the local nfvbench code (Git working copy) instead of the master branch
of nfvbench repository on OPNFV Gerrit.  This is useful for testing
during the development phases.

When this option is used, the version of the generator VM image is
extended with nfvbench development version number to be able to
distinguish the development images from the latest published image.

Change-Id: If21cbf2be59ff4f4110ebc157af0f934b98bddcb
Signed-off-by: Gwenael Lambrouin <gwenael.lambrouin@orange.com>