Set pylint version to 2.10.2
If we don't specify a version of pylint, jenkins jobs in gerrit
sometimes fail for a reason unrelated to the submitted change because a
newer version of pylint adding new checks is used for the first time
during the job run.  This is very annoying.

Let's stick to the latest version of pylint known to work with nfvbench,
and we will consider an update later on.

Change-Id: Ia2f87d0730ec1e733905a002a651ad6bcdf7edbb
Signed-off-by: Gwenael Lambrouin <gwenael.lambrouin@orange.com>