behave_tests: fix infinite recursion
Fix bug in behave_tests that occurs when behave looks for a previous
result (either non-regression or characterization) in testapi database.

When the result cannot be found on the first results page, behave enters
an infinite recursion.  Eventually, the operating system kills the
python interpreter leading to a return code equal to 137.

When behave is run by xtesting run_tests command, run_tests is also
killed so the error is not reported by xtesting.

This is now fixed.

Signed-off-by: Gwenael Lambrouin <gwenael.lambrouin@orange.com>
Change-Id: I6b0feafb5ebadf7d0d1df6d0ee03fd22cbe6899d