tox: only look for unit tests in test/
When the traffic generator VM is built with option -s (use local
nfvbench code), the unit test files are copied to
nfvbenchvm/dib/elements/nfvbenchvm/static/opt/nfvbench.  If pytest is
invoked after that, the test collection will fail because of duplicated
test files.

To fix this this issue that can occur on a developper's machine, we tell
pytest to look for test files in the test/ directory rather than
everywhere.  This is fine because test/ is the base directory for all
test files.

Change-Id: Idb7f62828db4290f16221bb72bd9c6d5d958f649
Signed-off-by: Gwenael Lambrouin <gwenael.lambrouin@orange.com>