tox: fix testing issues after VM build
When the traffic generator VM is built with option -s (use local
nfvbench code), the code and the unit test files are copied to
nfvbenchvm/dib/elements/nfvbenchvm/static/opt/nfvbench.  This can lead
to two problems:

1) if pytest is invoked after that, the test collection will fail
   because of duplicated test files;

2) if flake8 is invoked after that, the same files will be checked twice.

To fix this this issue that can occur on a developper's machine, we tell
pytest to look for test files in the test/ directory rather than
everywhere.  This is fine because test/ is the base directory for all
test files.

Similarly, we tell flake8 to skip the nfvbenchvm/ directory.

Change-Id: Idb7f62828db4290f16221bb72bd9c6d5d958f649
Signed-off-by: Gwenael Lambrouin <gwenael.lambrouin@orange.com>